+++ src/extension.ts
@@ -7,6 +7,7 @@ import { registerTerminalCommand } from './commands/terminal';
 import { registerLogCommand } from './commands/log';
 import { registerWatchCommand } from './commands/watch';
 import { registerVersionCommand } from './commands/version';
+import { registerUpdateCommand } from './commands/update';
 import { registerCheckoutCommand } from './commands/checkout';
 import { registerMergeCommand } from './commands/merge';
 import { registerDeleteCommand } from './commands/delete';
@@ -47,6 +48,7 @@ const activateExtension = async (context: vscode.ExtensionContext): Promise<void
     registerLogCommand(context, { extensionPath: context.extensionPath });
     registerWatchCommand(context, { extensionPath: context.extensionPath });
     registerVersionCommand(context);
+    registerUpdateCommand(context);
     registerCheckoutCommand(context);
     registerMergeCommand(context);
     registerDeleteCommand(context);